Template: vite-cf-DO-v2-runner
Base template for Create an app to help stop vaping. I've attached images of a UI from another app that helps stop vaping that I want you to base your UI off of. Create a catchy and relevant name, add rewards so the user feels inclined to keep opening the app and checking their progress. Now, with the UI images, don't directly copy them, as this is already another app. So I want you to just base your UI off of this current UI that I've attached. Ensure all features are working and statistics are working. Add a Google or email login so that the user, even if they switch devices or log out, can log back in and see their progress again. Keep in mind this app will be used on Android mobile devices, and the screen should fit nicely and adjust to any screen size that the user has for their Android mobile device.